backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
flask-pymongo==2.3.0
python-dotenv==1.0.0
werkzeug==3.1.3
pymongo==4.3.3
PyJWT==2.8.0
